+++ okhttp-clients/src/main/java/com/palantir/conjure/java/okhttp/OkhttpTraceInterceptor.java
@@ -42,18 +42,18 @@ public final class OkhttpTraceInterceptor implements Interceptor {
         } finally {
             // when we reach this point, we've got a 'Response' object (so the headers have come back), but the server
             // hasn't necessarily filled in the request body.
-            DetachedSpan waitForBody = chain.request().tag(AttemptSpan.class)
+            DetachedSpan waitForBody = chain.request().tag(Tags.AttemptSpan.class)
                     .attemptSpan()
                     .childDetachedSpan("OkHttp: wait-for-body", SpanType.CLIENT_OUTGOING);
 
-            chain.request().tag(SettableWaitForBodySpan.class).setWaitForBodySpan(waitForBody);
+            chain.request().tag(Tags.SettableWaitForBodySpan.class).setWaitForBodySpan(waitForBody);
         }
     }
 
     @SuppressWarnings("MustBeClosedChecker") // the OkhttpTraceInterceptor2 will definitely close this
     private static CloseableSpan createSpan(Request request) {
         return request
-                .tag(AttemptSpan.class)
+                .tag(Tags.AttemptSpan.class)
                 .attemptSpan()
                 .childSpan("OkHttp: wait-for-headers", SpanType.CLIENT_OUTGOING);
     }
