--- okhttp-clients/src/main/java/com/palantir/conjure/java/okhttp/SettableDispatcherSpan.java
@@ -1,32 +1,0 @@
-/*
- * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.conjure.java.okhttp;
-
-import com.palantir.conjure.java.client.config.ImmutablesStyle;
-import com.palantir.tracing.DetachedSpan;
-import org.immutables.value.Value;
-
-@Value.Modifiable
-@ImmutablesStyle
-interface SettableDispatcherSpan {
-    DetachedSpan dispatcherSpan();
-    SettableDispatcherSpan setDispatcherSpan(DetachedSpan span);
-
-    static SettableDispatcherSpan create() {
-        return ModifiableSettableDispatcherSpan.create();
-    }
-}
